+++ packages/web3-eth/test/integration/block/rpc.getBlock.test.ts
@@ -42,6 +42,8 @@ describe('rpc with block', () => {
 		earliest: 'earliest';
 		latest: 'latest';
 		pending: 'pending';
+		finalized: 'finalized';
+		safe: 'safe';
 		blockNumber: number | bigint;
 		blockHash: string;
 		transactionHash: string;
@@ -85,6 +87,8 @@ describe('rpc with block', () => {
 			pending: 'pending',
 			latest: 'latest',
 			earliest: 'earliest',
+			finalized: 'finalized',
+			safe: 'safe',
 			blockNumber: Number(receipt.blockNumber),
 			blockHash: String(receipt.blockHash),
 			transactionHash: String(receipt.transactionHash),
@@ -99,11 +103,18 @@ describe('rpc with block', () => {
 	describe('methods', () => {
 		it.each(
 			toAllVariants<{
-				block: 'earliest' | 'latest' | 'pending' | 'blockHash' | 'blockNumber';
+				block:
+					| 'earliest'
+					| 'latest'
+					| 'pending'
+					| 'finalized'
+					| 'safe'
+					| 'blockHash'
+					| 'blockNumber';
 				hydrated: boolean;
 				format: string;
 			}>({
-				block: ['earliest', 'latest', 'blockHash', 'blockNumber'],
+				block: ['earliest', 'latest', 'safe', 'finalized', 'blockHash', 'blockNumber'],
 				hydrated: [true, false],
 				format: Object.values(FMT_NUMBER),
 			}),